requirements.txt
altair==4.2.0
hydralit==1.0.12
hydralit-components==1.0.9

lxml==4.7.1
numpy==1.21.4
pandas==1.3.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
pyarrow==7.0.0
streamlit==1.5.1
streamlit-aggrid==0.2.3.post2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.8
